+++ core/core.pl
@@ -2505,7 +2505,7 @@ create_logtalk_flag(Flag, Value, Options) :-
 % versions, 'rcN' for release candidates (with N being a natural number),
 % and 'stable' for stable versions
 
-'$lgt_version_data'(logtalk(3, 2, 0, rc1)).
+'$lgt_version_data'(logtalk(3, 2, 0, rc2)).
 
 
 
@@ -3706,14 +3706,14 @@ create_logtalk_flag(Flag, Value, Options) :-
 
 
 
-% '$lgt_expand_term'(+object_identifier, ?term, ?term, +object_identifier, @scope)
+% '$lgt_expand_term'(+object_identifier, ?term, ?term, +object_identifier, @execution_context, @scope)
 %
 % expand_term/2 built-in method
 
-'$lgt_expand_term'(Obj, Term, Expansion, Sender, Scope) :-
+'$lgt_expand_term'(Obj, Term, Expansion, Sender, ExCtx, Scope) :-
 	(	var(Term) ->
 		Expansion = Term
-	;	'$lgt_term_expansion'(Obj, Term, Expand, Sender, Scope) ->
+	;	'$lgt_term_expansion'(Obj, Term, Expand, Sender, ExCtx, Scope) ->
 		Expansion = Expand
 	;	Term = (_ --> _) ->
 		% default grammar rule expansion
@@ -3738,7 +3738,7 @@ create_logtalk_flag(Flag, Value, Options) :-
 
 
 
-% '$lgt_term_expansion'(+object_identifier, ?term, ?term, +object_identifier, @scope)
+% '$lgt_term_expansion'(+object_identifier, ?term, ?term, +object_identifier, @execution_context, @scope)
 %
 % calls the term_expansion/2 user-defined predicate
 %
@@ -3746,19 +3746,20 @@ create_logtalk_flag(Flag, Value, Options) :-
 % when there is no scope directive, then we call any local definition when the sender
 % and the target object are the same
 
-'$lgt_term_expansion'(Obj, Term, Expansion, Sender, LookupScope) :-
+'$lgt_term_expansion'(Obj, Term, Expansion, Sender, ParentExCtx, LookupScope) :-
 	'$lgt_current_object_'(Obj, _, Dcl, Def, _, _, _, _, DDef, _, _),
 	(	call(Dcl, term_expansion(_, _), PredScope, _, _, SCtn, _) ->
 		(	(PredScope = LookupScope; Sender = SCtn) ->
-			'$lgt_execution_context'(ExCtx, _, Sender, Obj, Obj, _, _),
+			'$lgt_execution_context'(ExCtx, Obj, Sender, Obj, Obj, [], []),
 			call(Def, term_expansion(Term, Expansion), ExCtx, Call, _, _)
 		;	fail
 		)
-	;	Obj = Sender,
-		'$lgt_execution_context'(ExCtx, _, Obj, Obj, Obj, _, _),
-		(	call(Def, term_expansion(Term, Expansion), ExCtx, Call) ->
+	;	% no declaration for the term_expansion/2 hook predicate found
+		Obj = Sender,
+		% local call
+		(	call(Def, term_expansion(Term, Expansion), ParentExCtx, Call) ->
 			true
-		;	call(DDef, term_expansion(Term, Expansion), ExCtx, Call)
+		;	call(DDef, term_expansion(Term, Expansion), ParentExCtx, Call)
 		)
 	),
 	!,
@@ -3766,7 +3767,7 @@ create_logtalk_flag(Flag, Value, Options) :-
 
 
 
-% '$lgt_expand_goal'(+object_identifier, ?term, ?term, +object_identifier, @scope)
+% '$lgt_expand_goal'(+object_identifier, ?term, ?term, +object_identifier, @execution_context, @scope)
 %
 % expand_goal/2 built-in method
 %
@@ -3774,46 +3775,55 @@ create_logtalk_flag(Flag, Value, Options) :-
 % when there is no scope directive but the sender and the target objects are the
 % same, it calls any local definition of the goal_expansion/2 user-defined method
 
-'$lgt_expand_goal'(Obj, Goal, ExpandedGoal, Sender, LookupScope) :-
+'$lgt_expand_goal'(Obj, Goal, ExpandedGoal, Sender, ParentExCtx, LookupScope) :-
 	'$lgt_current_object_'(Obj, _, Dcl, Def, _, _, _, _, DDef, _, _),
-	(	call(Dcl, goal_expansion(_, _), PredScope, _, _, SCtn, _) ->
+	(	% lookup visible goal_expansion/2 hook predicate declaration
+		call(Dcl, goal_expansion(_, _), PredScope, _, _, SCtn, _) ->
 		(	(PredScope = LookupScope; Sender = SCtn) ->
-			'$lgt_execution_context'(ExCtx, _, Sender, Obj, Obj, _, _),
+			'$lgt_execution_context'(ExCtx, Obj, Sender, Obj, Obj, [], []),
 			'$lgt_expand_goal_scoped'(Goal, ExpandedGoal, Def, ExCtx)
 		;	ExpandedGoal = Goal
 		)
-	;	Obj = Sender ->
-		'$lgt_execution_context'(ExCtx, _, Obj, Obj, Obj, _, _),
-		'$lgt_expand_goal_local'(Goal, ExpandedGoal, Def, DDef, ExCtx)
-	;	ExpandedGoal = Goal
+	;	% no declaration for the goal_expansion/2 hook predicate found
+		Obj = Sender ->
+		% local call
+		'$lgt_expand_goal_local'(Goal, ExpandedGoal, Def, DDef, ParentExCtx)
+	;	% no goal_expansion/2 hook predicate definition within scope
+		ExpandedGoal = Goal
 	).
 
 
 '$lgt_expand_goal_scoped'(Goal, ExpandedGoal, Def, ExCtx) :-
 	(	var(Goal) ->
 		ExpandedGoal = Goal
-	;	call(Def, goal_expansion(Goal, ExpandedGoal0), ExCtx, Call, _, _) ->
+	;	% lookup visible goal_expansion/2 hook predicate definition
+		call(Def, goal_expansion(Goal, ExpandedGoal0), ExCtx, Call, _, _) ->
 		(	call(Call),
 			Goal \== ExpandedGoal0 ->
 			'$lgt_expand_goal_scoped'(ExpandedGoal0, ExpandedGoal, Def, ExCtx)
-		;	ExpandedGoal = Goal
+		;	% fixed-point found
+			ExpandedGoal = Goal
 		)
-	;	ExpandedGoal = Goal
+	;	% no visible goal_expansion/2 hook predicate definition found
+		ExpandedGoal = Goal
 	).
 
 
 '$lgt_expand_goal_local'(Goal, ExpandedGoal, Def, DDef, ExCtx) :-
 	(	var(Goal) ->
 		ExpandedGoal = Goal
-	;	(	call(Def, goal_expansion(Goal, ExpandedGoal0), ExCtx, Call)
+	;	% lookup local goal_expansion/2 hook predicate definition
+		(	call(Def, goal_expansion(Goal, ExpandedGoal0), ExCtx, Call)
 		;	call(DDef, goal_expansion(Goal, ExpandedGoal0), ExCtx, Call)
 		) ->
 		(	call(Call),
 			Goal \== ExpandedGoal0 ->
 			'$lgt_expand_goal_local'(ExpandedGoal0, ExpandedGoal, Def, DDef, ExCtx)
-		;	ExpandedGoal = Goal
+		;	% fixed-point found
+			ExpandedGoal = Goal
 		)
-	;	ExpandedGoal = Goal
+	;	% no local goal_expansion/2 hook predicate definition found
+		ExpandedGoal = Goal
 	).
 
 
@@ -10493,13 +10503,13 @@ create_logtalk_flag(Flag, Value, Options) :-
 	!,
 	'$lgt_comp_ctx'(Ctx, _, Entity, _, This, _, _, _, _, ExCtx, _, _, _),
 	'$lgt_db_call_database_execution_context'(Entity, This, Database, ExCtx),
-	TPred = '$lgt_expand_term'(Database, Term, Expansion, Database, p(_)).
+	TPred = '$lgt_expand_term'(Database, Term, Expansion, Database, ExCtx, p(_)).
 
 '$lgt_compile_body'(expand_goal(Goal, ExpandedGoal), TPred, '$lgt_debug'(goal(expand_goal(Goal, ExpandedGoal), TPred), ExCtx), Ctx) :-
 	!,
 	'$lgt_comp_ctx'(Ctx, _, Entity, _, This, _, _, _, _, ExCtx, _, _, _),
 	'$lgt_db_call_database_execution_context'(Entity, This, Database, ExCtx),
-	TPred = '$lgt_expand_goal'(Database, Goal, ExpandedGoal, Database, p(_)).
+	TPred = '$lgt_expand_goal'(Database, Goal, ExpandedGoal, Database, ExCtx, p(_)).
 
 % DCG predicates
 
@@ -12046,12 +12056,12 @@ create_logtalk_flag(Flag, Value, Options) :-
 
 % term and goal expansion predicates
 
-'$lgt_compile_message_to_object'(expand_term(Term, Expansion), Obj, '$lgt_expand_term'(Obj, Term, Expansion, This, p(p(p))), _, Ctx) :-
+'$lgt_compile_message_to_object'(expand_term(Term, Expansion), Obj, '$lgt_expand_term'(Obj, Term, Expansion, This, ExCtx, p(p(p))), _, Ctx) :-
 	!,
 	'$lgt_comp_ctx'(Ctx, _, _, _, This, _, _, _, _, ExCtx, _, _, _),
 	'$lgt_execution_context_this_entity'(ExCtx, This, _).
 
-'$lgt_compile_message_to_object'(expand_goal(Goal, ExpandedGoal), Obj, '$lgt_expand_goal'(Obj, Goal, ExpandedGoal, This, p(p(p))), _, Ctx) :-
+'$lgt_compile_message_to_object'(expand_goal(Goal, ExpandedGoal), Obj, '$lgt_expand_goal'(Obj, Goal, ExpandedGoal, This, ExCtx, p(p(p))), _, Ctx) :-
 	!,
 	'$lgt_comp_ctx'(Ctx, _, _, _, This, _, _, _, _, ExCtx, _, _, _),
 	'$lgt_execution_context_this_entity'(ExCtx, This, _).
@@ -12282,15 +12292,15 @@ create_logtalk_flag(Flag, Value, Options) :-
 
 % term and goal expansion predicates
 
-'$lgt_compile_message_to_self'(expand_term(Term, Expansion), '$lgt_expand_term'(Self, Term, Expansion, This, p(_)), Ctx) :-
+'$lgt_compile_message_to_self'(expand_term(Term, Expansion), '$lgt_expand_term'(Self, Term, Expansion, This, ExCtx, p(_)), Ctx) :-
 	!,
-	'$lgt_comp_ctx_self'(Ctx, Self),
-	'$lgt_comp_ctx_this'(Ctx, This).
+	'$lgt_comp_ctx'(Ctx, _, _, _, This, Self, _, _, _, ExCtx, _, _, _),
+	'$lgt_execution_context'(ExCtx, _, _, This, Self, _, _).
 
-'$lgt_compile_message_to_self'(expand_goal(Goal, ExpandedGoal), '$lgt_expand_goal'(Self, Goal, ExpandedGoal, This, p(_)), Ctx) :-
+'$lgt_compile_message_to_self'(expand_goal(Goal, ExpandedGoal), '$lgt_expand_goal'(Self, Goal, ExpandedGoal, ExCtx, p(_)), Ctx) :-
 	!,
-	'$lgt_comp_ctx_self'(Ctx, Self),
-	'$lgt_comp_ctx_this'(Ctx, This).
+	'$lgt_comp_ctx'(Ctx, _, _, _, This, Self, _, _, _, ExCtx, _, _, _),
+	'$lgt_execution_context'(ExCtx, _, _, This, Self, _, _).
 
 % compiler bypass control construct